chore: move integration related logic closer together
