+++ wandb/integration/openai/resolver.py
@@ -5,8 +5,8 @@ from dataclasses import asdict, dataclass
 from typing import Any, Dict, List, Optional, Sequence
 
 import wandb
+from wandb.integration.integration_utils.auto_logging import Response
 from wandb.sdk.data_types import trace_tree
-from wandb.sdk.integration_utils.auto_logging import Response
 
 logger = logging.getLogger(__name__)
